+++ src/Components/SupportsExtraDbConfigs.php
@@ -1,9 +1,10 @@
 <?php
+
 namespace DreamFactory\Core\Database\Components;
 
 use DreamFactory\Core\Models\ServiceCacheConfig;
 
-trait SupportsUpsertAndCache
+trait SupportsExtraDbConfigs
 {
     /**
      * {@inheritdoc}
@@ -16,6 +17,7 @@ trait SupportsUpsertAndCache
         }
 
         $config['allow_upsert'] = filter_var(array_get($local_config, 'allow_upsert'), FILTER_VALIDATE_BOOLEAN);
+        $config['max_records'] = filter_var(array_get($local_config, 'max_records'), FILTER_VALIDATE_INT);
 
         return $config;
     }
@@ -30,7 +32,14 @@ trait SupportsUpsertAndCache
         $result = (array)parent::setConfig($id, $config, $local_config);
 
         if (isset($config['allow_upsert'])) {
-            $result['allow_upsert'] = filter_var(array_get($config, 'allow_upsert'), FILTER_VALIDATE_BOOLEAN);
+            $result['allow_upsert'] = filter_var($config['allow_upsert'], FILTER_VALIDATE_BOOLEAN);
+        }
+
+        if (isset($config['max_records'])) {
+            $result['max_records'] = filter_var($config['max_records'], FILTER_VALIDATE_INT);
+        }
+        if (array_get($result, 'max_records', 0) <= 0 || empty($config['max_records'])) {
+            unset($result['max_records']);
         }
 
         return $result;
@@ -59,6 +68,14 @@ trait SupportsUpsertAndCache
             'allow_null'  => false,
             'default'     => false,
             'description' => 'Allow PUT to create records if they do not exist and the service is capable.',
+        ];
+        $schema[] = [
+            'name'        => 'max_records',
+            'label'       => 'Maximum Records',
+            'type'        => 'integer',
+            'allow_null'  => false,
+            'default'     => 1000,
+            'description' => 'Maximum number of records returned by this service. Must be a number greater than 0. Default is 1000.',
         ];
         $schema = array_merge($schema, ServiceCacheConfig::getConfigSchema());